fix groudtruthcomparison
